feat(xmonad): add functionality
Altered the workspace generated string (outputted to xmobar).  I
changed the names of the workspaces, but ultimately stripped the
string names and removed the font color to produce minimal boxes
to indicate active / inactive workspaces.  Also, I added a Logger
to display the name of the current workspace to the right of the
aforementioned, minimal boxes :)