feat(nvim): migrate from minpac to packer
Packer leverages vim packages and provides a declarative api to manage
packages.  A huge bonus: written in Lua, which is a first class language
of Neovim.

Additionally, removed tmux-navigator-vim in favor of a lua alternative.
I'm not sure if I'm actually going to keep this plugin at all.  We'll
see if my workflow is bothered in lieu of its absense.