feat(nvim): new plug keymap, remove dirvish opts
We migrated away from dirvish, in favor of the `lir` (tamago324's lua
file manager); so, naturally, we remove all config and keys mapped to
dirvish.  We leave one option that was lumped with dirvish config: netrw
disable.

I liked the keybind that dirvish provided out of the box ('-' to file
explore relative to the current buffer).  So, we added that keybind
under lir the config.